third_party/WebKit/LayoutTests/typedcssom/cssNumberValue.html

<!DOCTYPE html>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>

<script>
var values = [
  {input: new CSSNumberValue(0), cssString: '0'},
  {input: new CSSNumberValue(1), cssString: '1'},
  {input: new CSSNumberValue(-2), cssString: '-2'},
  {input: new CSSNumberValue(3.4), cssString: '3.4'}
];

test(function() {
  for (var i = 0; i < values.length; ++i) {
    assert_equals(values[i].input.cssString, values[i].cssString);
  }
}, "Test that the css string for CSSNumberValue is correct.");

</script>

<body>
</body>
